refactor: Rename typography file to index for easier import
