+++ src/components/typography/caption/caption.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from "react";
-import { typography } from "../../../typography/typography";
+import { typography } from "../../../typography";
 
 function Caption({ children }: { children: ReactNode }) {
   return <caption style={typography.caption}>{children}</caption>;
