+++ src/components/typography/details/details.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from "react";
-import { typography } from "../../../typography/typography";
+import { typography } from "../../../typography";
 
 function Details({ children }: { children: ReactNode }) {
   return <details style={typography.details}>{children}</details>;
