+++ src/components/typography/display/display.tsx
@@ -1,8 +1,7 @@
 import { ReactNode } from "react";
-import { typography } from "../../../typography/typography";
+import { typography } from "../../../typography";
 
 function Display({ children }: { children: ReactNode }) {
-    return <p style={typography.display}>{children}</p>;
-    
+  return <p style={typography.display}>{children}</p>;
 }
 export default Display;
