+++ src/components/typography/headline/h1.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from "react";
-import { typography } from "../../../typography/typography";
+import { typography } from "../../../typography";
 
 function Heading1({ children }: { children: ReactNode }) {
   return <h1 style={typography.H1}>{children}</h1>;
