+++ src/components/typography/headline/h2.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from "react";
-import { typography } from "../../../typography/typography";
+import { typography } from "../../../typography";
 
 function Heading2({ children }: { children: ReactNode }) {
   return <h2 style={typography.H2}>{children}</h2>;
