+++ src/components/typography/headline/h3.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from "react";
-import { typography } from "../../../typography/typography";
+import { typography } from "../../../typography";
 
 function Heading3({ children }: { children: ReactNode }) {
   return <h3 style={typography.H3}>{children}</h3>;
