+++ src/components/typography/headline/h4.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from "react";
-import { typography } from "../../../typography/typography";
+import { typography } from "../../../typography";
 
 function Heading4({ children }: { children: ReactNode }) {
   return <h4 style={typography.H4}>{children}</h4>;
