+++ src/components/typography/label/labelLargeBold.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from "react";
-import { typography } from "../../../typography/typography";
+import { typography } from "../../../typography";
 
 function LabelLargeBold({ children }: { children: ReactNode }) {
   return <label style={typography.label.labelLargeBold}>{children}</label>;
