+++ src/components/typography/label/labelLargeRegular.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from "react";
-import { typography } from "../../../typography/typography";
+import { typography } from "../../../typography";
 
 function LabelLargeRegular({ children }: { children: ReactNode }) {
   return <label style={typography.label.labelLargeRegular}>{children}</label>;
