+++ src/components/typography/label/largeLabel.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from "react";
-import { typography } from "../../../typography/typography";
+import { typography } from "../../../typography";
 
 function LargeLabel({ children }: { children: ReactNode }) {
   return <label style={typography.label.largeLabel}>{children}</label>;
