+++ src/components/typography/price/bigPrice.tsx
@@ -1,8 +1,7 @@
 import { ReactNode } from "react";
-import { typography } from "../../../typography/typography";
+import { typography } from "../../../typography";
 
 function BigPrice({ children }: { children: ReactNode }) {
-    return <p style={typography.price.bigPrice}>{children}</p>;
-    
+  return <p style={typography.price.bigPrice}>{children}</p>;
 }
 export default BigPrice;
