+++ src/components/typography/price/mediumPrice.tsx
@@ -1,8 +1,7 @@
 import { ReactNode } from "react";
-import { typography } from "../../../typography/typography";
+import { typography } from "../../../typography";
 
 function MediumPrice({ children }: { children: ReactNode }) {
-    return <p style={typography.price.mediumPrice}>{children}</p>;
-    
+  return <p style={typography.price.mediumPrice}>{children}</p>;
 }
 export default MediumPrice;
