+++ src/components/typography/price/smallPrice.tsx
@@ -1,8 +1,7 @@
 import { ReactNode } from "react";
-import { typography } from "../../../typography/typography";
+import { typography } from "../../../typography";
 
 function SmallPrice({ children }: { children: ReactNode }) {
-    return <p style={typography.price.smallPrice}>{children}</p>;
-    
+  return <p style={typography.price.smallPrice}>{children}</p>;
 }
 export default SmallPrice;
